# openapi.yaml
openapi: 3.1.0
info:
  title: RAIDHack API
  description: |
    RAIDHackは、開発者がアイデアを共有し、ハッカソンをするためのプラットフォームです。
  version: 1.0.0
  contact:
    name: RAIDHack Platform
    url: https://raidhack-web.pages.dev
servers:
  - url: https://raidhack-api.ukawamochi5.workers.dev
    description: APIサーバー
paths:
...
  /api/notifications:
    get:
      tags:
        - Notifications
      summary: 通知一覧取得
      description: 認証済みユーザーの通知一覧を取得します
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: ページ番号
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: 1ページあたりの件数
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: unread_only
          in: query
          description: 未読のみ取得
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: 通知一覧取得成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications/unread-count:
    get:
      tags:
        - Notifications
      summary: 未読通知数取得
      description: 認証済みユーザーの未読通知数を取得します
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 未読通知数取得成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      count:
                        type: integer
                        example: 3
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications/{id}/read:
    put:
      tags:
        - Notifications
      summary: 通知既読マーク
      description: 指定された通知を既読にします
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: 既読処理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/notifications/read-all:
    put:
      tags:
        - Notifications
      summary: 全通知既読マーク
      description: 認証済みユーザーの全ての通知を既読にします
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 全既読処理成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/notifications/{id}:
    delete:
      tags:
        - Notifications
      summary: 通知削除
      description: 指定された通知を削除します
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: 通知削除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===== プロジェクト関連 =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: GitHub OAuth認証で取得したJWTトークンをAuthorizationヘッダーに含めてください
  parameters:
...
    NotificationId:
      name: id
      in: path
      required: true
      description: 通知ID
      schema:
        type: integer
        example: 1
  responses:
...
    Unauthorized:
      description: 認証が必要です
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: リソースが見つかりません
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    SuccessResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "操作が正常に完了しました"
    ErrorResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "エラーが発生しました"
        error:
          type: string
          example: "詳細なエラーメッセージ"
...
    Notification:
      type: object
      required:
        - id
        - user_id
        - type
        - title
        - message
        - is_read
        - created_at
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        type:
          type: string
          enum: [application, team_invite, application_status, new_idea, vote, system]
          example: "application"
        title:
          type: string
          example: "新しい応募があります"
        message:
          type: string
          example: "あなたのアイデア「タスク管理アプリ」に新しい応募があります"
        data:
          type: object
          nullable: true
          example: {"ideaId": 1, "applicationId": 2}
        is_read:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5

    # プロジェクト関連のスキーマ
...